+++ src/components/input/InputItem.tsx
@@ -46,12 +46,13 @@ const InputItem = forwardRef<HTMLInputElement, InputProps>(
             )}
           </>
         ) : (
-          <>
+          <div className="relative">
             <input
               className={cls(
                 "h-[50px] w-full rounded-lg px-4 text-lg text-black03 placeholder-gray02 ring-1 ring-inset ring-gray03 transition-all focus:outline-none focus:ring-1 focus:ring-inset",
                 errors ? "ring-red01 transition-all focus-within:ring-red01" : "focus-within:ring-violet01",
-                readOnly ? "bg-gray03" : ""
+                readOnly ? "bg-gray03" : "",
+                type === "password" ? "pr-10" : ""
               )}
               id={id}
               name={name}
@@ -62,7 +63,7 @@ const InputItem = forwardRef<HTMLInputElement, InputProps>(
               {...props}
             />
             {type === "password" && !readOnly && (
-              <span className="absolute right-3 top-[47px] cursor-pointer" onClick={handleIconClick}>
+              <span className="absolute right-3 top-1/2 -translate-y-1/2 cursor-pointer" onClick={handleIconClick}>
                 {password ? (
                   <Image src={OpenEyes} width={24} height={24} alt="OpenEyes" />
                 ) : (
@@ -70,7 +71,7 @@ const InputItem = forwardRef<HTMLInputElement, InputProps>(
                 )}
               </span>
             )}
-          </>
+          </div>
         )}
 
         {errors && <span className="pl-1 text-base font-normal text-red01">{errors}</span>}
